- project will be applied only to the root project - initialization of provider changed to reduce calculation of properties - implementation-titel removed, because this is not longer possible for subprojects
