+++ src/main/groovy/com/intershop/gradle/buildinfo/scm/GitScmInfoProvider.groovy
@@ -36,6 +36,11 @@ class GitScmInfoProvider extends AbstractScmInfoProvider {
      */
     private final Repository gitRepo
 
+    private String pOrigin = ''
+    private String pBranchName = ''
+    private String pRevInfo = ''
+    private String pChangeTime = ''
+
     /**
      * Constructs the Git information provider
      * @param projectDir
@@ -51,21 +56,22 @@ class GitScmInfoProvider extends AbstractScmInfoProvider {
      */
     @Override
     String getSCMOrigin() {
-        String origin = ''
-        if(bambooBuild) {
-            origin = System.getenv('bamboo_planRepository_repositoryUrl')
-        }
-        if(! origin) {
-            Config config = gitRepo.config
-            String rv = config.getString('remote', 'origin', 'url')
+        if(! pOrigin) {
+            if (bambooBuild) {
+                pOrigin = System.getenv('bamboo_planRepository_repositoryUrl')
+            }
+            if (! pOrigin) {
+                Config config = gitRepo.config
+                String rv = config.getString('remote', 'origin', 'url')
 
-            if (rv && rv.startsWith('https://') && rv.contains('@')) {
-                origin = "https://${rv.substring(rv.indexOf('@') + 1)}"
-            } else {
-                origin = rv
+                if (rv && rv.startsWith('https://') && rv.contains('@')) {
+                    pOrigin = "https://${rv.substring(rv.indexOf('@') + 1)}"
+                } else {
+                    pOrigin = rv ? rv : UNKNOWN
+                }
             }
         }
-        return origin
+        return pOrigin
     }
 
     /**
@@ -74,22 +80,26 @@ class GitScmInfoProvider extends AbstractScmInfoProvider {
      */
     @Override
     String getBranchName() {
-        String rv = gitRepo.branch
+        if(! pBranchName) {
+            String rv = gitRepo.branch
 
-        Git git = new Git(gitRepo)
-        List<Ref> refList = git.tagList().call()
+            Git git = new Git(gitRepo)
+            List<Ref> refList = git.tagList().call()
 
-        refList.any {Ref ref ->
-            Ref peeledRef = gitRepo.peel(ref)
-            String hashID =  ref.getObjectId()
-            hashID = peeledRef.getPeeledObjectId()?.getName()
-            if(hashID == rv) {
-                rv = ref.getName()
-                return true
+            refList.any {Ref ref ->
+                Ref peeledRef = gitRepo.peel(ref)
+                String hashID =  ref.getObjectId()
+                hashID = peeledRef.getPeeledObjectId()?.getName()
+                if(hashID == rv) {
+                    rv = ref.getName()
+                    return true
+                }
             }
+
+            pBranchName = rv.split('/').last()
         }
 
-        return rv.split('/').last()
+        return pBranchName
     }
 
     /**
@@ -98,14 +108,17 @@ class GitScmInfoProvider extends AbstractScmInfoProvider {
      */
     @Override
     String getSCMRevInfo() {
-        ObjectId id = gitRepo.resolve(Constants.HEAD)
-        String rv = ''
+        if(! pRevInfo) {
+            ObjectId id = gitRepo.resolve(Constants.HEAD)
+            String rv = ''
 
-        if(id) {
-            rv = id.name?.substring(0,7)
+            if (id) {
+                pRevInfo = id.name?.substring(0, 7)
+            } else {
+                pRevInfo = UNKNOWN
+            }
         }
-
-        return rv
+        return pRevInfo
     }
 
     /**
@@ -114,12 +127,14 @@ class GitScmInfoProvider extends AbstractScmInfoProvider {
      */
     @Override
     String getLastChangeTime() {
-        if(SCMRevInfo) {
+        if(! getSCMRevInfo().equals('unknown')) {
             RevWalk walk = new RevWalk(gitRepo)
             RevCommit commit = walk.parseCommit(gitRepo.resolve(Constants.HEAD))
             return new Date( ((long)commit.commitTime)*1000).format("yyyyMMddHHmmss")
+        } else {
+            pChangeTime = UNKNOWN
         }
-        return ''
+        return pChangeTime
     }
 
     /**
