+++ src/test/groovy/com/intershop/gradle/buildinfo/LocalBuildInfoSpec.groovy
@@ -74,8 +74,6 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         (new File(testProjectDir, 'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml')).exists()
 
         String ivyFileContents = new File(testProjectDir,  'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml').text
-        assertTrue(ivyFileContents.contains('com.test:test:1.0.0.0'))
-
         1 == ivyFileContents.count('<e:scm-type>local</e:scm-type>')
 
         when:
@@ -87,12 +85,10 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         then:
         secResult.output.contains('Add buildinfo to ivy file')
         ! secResult.output.contains('Add buildinfo to manifest')
-        ! result.output.contains('Buildinfo for manifest will be extended!')
+        result.output.contains('Buildinfo for manifest will be extended!')
         (new File(testProjectDir, 'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml')).exists()
 
         String secIvyFileContents = new File(testProjectDir,  'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml').text
-        assertTrue(secIvyFileContents.contains('com.test:test:1.0.0.0'))
-
         1 == secIvyFileContents.count('<e:scm-type>local</e:scm-type>')
 
         where:
@@ -148,8 +144,6 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         (new File(testProjectDir, 'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml')).exists()
 
         String ivyFileContents = new File(testProjectDir,  'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml').text
-        assertTrue(ivyFileContents.contains('com.test:test:1.0.0.0'))
-
         1 == ivyFileContents.count('<e:scm-type>local</e:scm-type>')
 
         where:
@@ -202,16 +196,13 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         (new File(testProjectDir, 'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml')).exists()
         (new File(testProjectDir, 'build/tmp/jar/MANIFEST.MF')).exists()
 
-        String mfFileContents = new File(testProjectDir, 'build/tmp/jar/MANIFEST.MF').text
-        assertTrue(mfFileContents.contains('Implementation-Title: com.test:test:1.0.0.0'))
-
         String ivyFileContents = new File(testProjectDir,  'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml').text
         assertFalse(ivyFileContents.contains('com.test:test:1.0.0.0'))
 
         0 == ivyFileContents.count('<e:scm-type>local</e:scm-type>')
 
         when:
-        def secResult = getPreparedGradleRunner()
+        getPreparedGradleRunner()
                 .withArguments('publish', '-PrunOnCI=true', '-PnoDescriptorInfo=true', '--stacktrace', '-i')
                 .withGradleVersion(gradleVersion)
                 .build()
@@ -220,8 +211,6 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         (new File(testProjectDir, 'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml')).exists()
 
         String secIvyFileContents = new File(testProjectDir,  'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml').text
-        assertFalse(secIvyFileContents.contains('com.test:test:1.0.0.0'))
-
         0 == secIvyFileContents.count('<e:scm-type>local</e:scm-type>')
 
         where:
@@ -265,7 +254,7 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         """.stripIndent()
 
         when:
-        def result = getPreparedGradleRunner()
+        getPreparedGradleRunner()
                 .withArguments('publish', '-PrunOnCI=true', '-PnoDescriptorInfo=true', '--stacktrace', '-i')
                 .withGradleVersion(gradleVersion)
                 .build()
@@ -274,16 +263,11 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         (new File(testProjectDir, 'build/repo/com/test/test/1.0.0.0/test-1.0.0.0.pom')).exists()
         (new File(testProjectDir, 'build/tmp/jar/MANIFEST.MF')).exists()
 
-        String mfFileContents = new File(testProjectDir, 'build/tmp/jar/MANIFEST.MF').text
-        assertTrue(mfFileContents.contains('Implementation-Title: com.test:test:1.0.0.0'))
-
         String pomFileContents = new File(testProjectDir, 'build/repo/com/test/test/1.0.0.0/test-1.0.0.0.pom').text
-        assertFalse(pomFileContents.contains('com.test:test:1.0.0.0'))
-
         0 == pomFileContents.count('<scm-type>local</scm-type>')
 
         when:
-        def secResult = getPreparedGradleRunner()
+        getPreparedGradleRunner()
                 .withArguments('publish', '-PrunOnCI=true', '-PnoDescriptorInfo=true', '--stacktrace', '-i')
                 .withGradleVersion(gradleVersion)
                 .build()
@@ -292,8 +276,6 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         (new File(testProjectDir, 'build/repo/com/test/test/1.0.0.0/test-1.0.0.0.pom')).exists()
 
         String secPomFileContents = new File(testProjectDir, 'build/repo/com/test/test/1.0.0.0/test-1.0.0.0.pom').text
-        assertFalse(secPomFileContents.contains('com.test:test:1.0.0.0'))
-
         0 == secPomFileContents.count('<scm-type>local</scm-type>')
 
         where:
@@ -346,12 +328,7 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         (new File(testProjectDir, 'build/repo/com/test/test/1.0.0.0/test-1.0.0.0.pom')).exists()
         (new File(testProjectDir, 'build/tmp/jar/MANIFEST.MF')).exists()
 
-        String mfFileContents = new File(testProjectDir, 'build/tmp/jar/MANIFEST.MF').text
-        assertTrue(mfFileContents.contains('Implementation-Title: com.test:test:1.0.0.0'))
-
         String pomFileContents = new File(testProjectDir, 'build/repo/com/test/test/1.0.0.0/test-1.0.0.0.pom').text
-        assertTrue(pomFileContents.contains('com.test:test:1.0.0.0'))
-
         1 == pomFileContents.count('<scm-type>local</scm-type>')
 
         when:
@@ -364,8 +341,6 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         (new File(testProjectDir, 'build/repo/com/test/test/1.0.0.0/test-1.0.0.0.pom')).exists()
 
         String secPomFileContents = new File(testProjectDir, 'build/repo/com/test/test/1.0.0.0/test-1.0.0.0.pom').text
-        assertTrue(secPomFileContents.contains('com.test:test:1.0.0.0'))
-
         1 == secPomFileContents.count('<scm-type>local</scm-type>')
 
         where:
@@ -414,8 +389,8 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         then:
         ! result.output.contains('Add buildinfo to pom file')
         ! result.output.contains('Add buildinfo to ivy file')
-        ! result.output.contains('Buildinfo for manifest will be extended!')
-        result.output.contains('Add buildinfo to manifest')
+        result.output.contains('Buildinfo for manifest will be extended!')
+        ! result.output.contains('Buildinfo for manifest will be extended with additional parameters.')
         (new File(testProjectDir, 'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml')).exists()
 
         String ivyFileContents = new File(testProjectDir, 'build/repo/com.test/test/1.0.0.0/ivy-1.0.0.0.xml').text
@@ -425,6 +400,110 @@ class LocalBuildInfoSpec extends AbstractIntegrationSpec {
         gradleVersion << supportedGradleVersions
     }
 
+    def 'Add BuildInfo to Jars in multi projects (Gradle #gradleVersion)'(gradleVersion) {
 
+        String prja = """
+                apply plugin: 'java'
+                apply plugin: 'ivy-publish'
+                group = 'com.intershop.testproject'
+                version = rootProject.getVersion()
+                publishing {
+                    repositories {
+                        ivy {
+                            // change to point to your repo, e.g. http://my.org/repo
+                            url "\${rootProject.buildDir}/repo"
+                        }
+                    }
+                    publications {
+                        ivy(IvyPublication) {
+                            from components.java
+                        }
+                    }
+                }
+                """.stripIndent()
+        String prjb = """
+                apply plugin: 'java'
+                apply plugin: 'ivy-publish'
+                group = 'com.intershop.testproject'
+                version = rootProject.getVersion()
+                publishing {
+                    repositories {
+                        ivy {
+                            // change to point to your repo, e.g. http://my.org/repo
+                            url "\${rootProject.buildDir}/repo"
+                        }
+                    }
+                    publications {
+                        ivy(IvyPublication) {
+                            from components.java
+                        }
+                    }
+                }
+                """.stripIndent()
+
+        File settingsfile = file('settings.gradle')
+        settingsfile << """
+            rootProject.name= 'multiprojecttest'
+        """.stripIndent()
+
+        buildFile << """
+            plugins {
+                id 'java'
+                id 'com.intershop.gradle.buildinfo'
+                id 'ivy-publish'
+            }
+            group = 'com.test.root'
+            version = '1.0.0'
+            sourceCompatibility = 1.7
+            targetCompatibility = 1.7
+            publishing {
+                repositories {
+                    ivy {
+                        // change to point to your repo, e.g. http://my.org/repo
+                        url "\${rootProject.buildDir}/repo"
+                    }
+                }
+                publications {
+                    ivy(IvyPublication) {
+                        from components.java
+                    }
+                }
+            }
+        """.stripIndent()
+
+        createSubProjectJava('project1a', settingsfile, 'com.intereshop.a', prja, '1.0.0')
+        createSubProjectJava('project2b', settingsfile, 'com.intereshop.b', prjb, '1.0.0')
+
+        when:
+        def result = getPreparedGradleRunner()
+                .withArguments('publish', '-PrunOnCI=true', '--stacktrace', '-i')
+                .withGradleVersion(gradleVersion)
+                .build()
+
+        then:
+        result.output.contains('Add buildinfo to ivy file')
+        result.output.contains('Add buildinfo to manifest')
+
+        (new File(testProjectDir, 'build/repo/com.test.root/multiprojecttest/1.0.0/ivy-1.0.0.xml')).exists()
+        (new File(testProjectDir, 'build/repo/com.intershop.testproject/project1a/1.0.0/ivy-1.0.0.xml')).exists()
+        (new File(testProjectDir, 'build/repo/com.intershop.testproject/project2b/1.0.0/ivy-1.0.0.xml')).exists()
+
+        String ivyFileContentsRoot = new File(testProjectDir,  'build/repo/com.test.root/multiprojecttest/1.0.0/ivy-1.0.0.xml').text
+        String ivyFileContentsA = new File(testProjectDir,  'build/repo/com.intershop.testproject/project1a/1.0.0/ivy-1.0.0.xml').text
+        String ivyFileContentsB = new File(testProjectDir,  'build/repo/com.intershop.testproject/project2b/1.0.0/ivy-1.0.0.xml').text
+
+        1 == ivyFileContentsRoot.count('<e:scm-type>local</e:scm-type>')
+        1 == ivyFileContentsA.count('<e:scm-type>local</e:scm-type>')
+        1 == ivyFileContentsB.count('<e:scm-type>local</e:scm-type>')
+
+        where:
+        gradleVersion << supportedGradleVersions
+    }
+
+    private File createSubProjectJava(String projectPath, File settingsGradle, String packageName, String buildContent, String version){
+        File subProject = createSubProject(projectPath, settingsGradle, buildContent)
+        writeJavaTestClass(packageName, subProject)
+        return subProject
+    }
 }
